requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking
#
anyio==3.6.2
    # via jupyter-server
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
argon2-cffi==21.3.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asciitree==0.3.3
    # via zarr
ase==3.22.1
    # via nexusutils
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via jsonschema
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   dask
    #   nexusutils
cloudpickle==2.2.1
    # via dask
comm==0.1.2
    # via ipykernel
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dask[array]==2023.1.0
    # via hyperspy
debugpy==1.6.5
    # via ipykernel
decorator==5.1.1
    # via
    #   ipyparallel
    #   ipython
defusedxml==0.7.1
    # via nbconvert
dill==0.3.6
    # via hyperspy
entrypoints==0.4
    # via
    #   ipyparallel
    #   jupyter-client
    #   numcodecs
executing==1.2.0
    # via stack-data
fasteners==0.18
    # via zarr
fastjsonschema==2.16.2
    # via nbformat
flatdict==4.0.1
    # via nexusutils
fonttools==4.38.0
    # via matplotlib
fsspec==2023.1.0
    # via
    #   dask
    #   hyperspy
future==0.18.3
    # via h5py-wrapper
h5grove==1.2.0
    # via jupyterlab-h5web
h5py==3.7.0
    # via
    #   h5grove
    #   h5py-wrapper
    #   hyperspy
    #   jupyterlab-h5web
    #   nexusutils
    #   pyelli
h5py-wrapper==1.1.0
    # via hdfdict
hdfdict @ git+https://github.com/shabihsherjeel/hdfdict@f48a586
    # via nexusutils
hyperspy==1.7.3
    # via nexusutils
idna==3.4
    # via
    #   anyio
    #   requests
imageio==2.24.0
    # via
    #   hyperspy
    #   scikit-image
importlib-metadata==6.0.0
    # via
    #   hyperspy
    #   nbconvert
    #   numba
importlib-resources==5.10.2
    # via
    #   jsonschema
    #   pyelli
ipykernel==6.20.2
    # via ipyparallel
ipyparallel==8.4.1
    # via hyperspy
ipython==8.8.0
    # via
    #   hyperspy
    #   ipykernel
    #   ipyparallel
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   hyperspy
    #   jupyter-server
    #   nbconvert
jsonschema==4.17.3
    # via nbformat
jupyter-client==7.4.9
    # via
    #   ipykernel
    #   ipyparallel
    #   jupyter-server
    #   nbclient
jupyter-core==5.1.3
    # via
    #   jupyter-client
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-server==1.23.5
    # via jupyterlab-h5web
jupyterlab-h5web==6.5.0
    # via nexusutils
jupyterlab-pygments==0.2.2
    # via nbconvert
kiwisolver==1.4.4
    # via matplotlib
llvmlite==0.39.1
    # via numba
locket==1.0.0
    # via partd
markupsafe==2.1.2
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.6.3
    # via
    #   ase
    #   hyperspy
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.4
    # via nbconvert
mpmath==1.2.1
    # via sympy
natsort==8.2.0
    # via hyperspy
nbclient==0.7.2
    # via nbconvert
nbconvert==7.2.8
    # via jupyter-server
nbformat==5.7.3
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
networkx==3.0
    # via scikit-image
nexusutils @ git+https://github.com/nomad-coe/nomad-parser-nexus.git@rii-reader
    # via -r requirements.in
numba==0.56.4
    # via
    #   hyperspy
    #   sparse
numcodecs==0.11.0
    # via zarr
numexpr==2.8.4
    # via hyperspy
numpy==1.23.5
    # via
    #   ase
    #   contourpy
    #   dask
    #   h5grove
    #   h5py
    #   hyperspy
    #   imageio
    #   matplotlib
    #   nexusutils
    #   numba
    #   numcodecs
    #   numexpr
    #   pandas
    #   pyelli
    #   pywavelets
    #   scikit-image
    #   scipy
    #   sparse
    #   tifffile
    #   xarray
    #   zarr
orjson==3.8.5
    # via h5grove
packaging==23.0
    # via
    #   dask
    #   hyperspy
    #   ipykernel
    #   jupyter-server
    #   matplotlib
    #   nbconvert
    #   scikit-image
    #   xarray
pandas==1.5.3
    # via
    #   -r requirements.in
    #   nexusutils
    #   pyelli
    #   xarray
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
partd==1.3.0
    # via dask
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.4.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
pint==0.20.1
    # via hyperspy
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==2.6.2
    # via jupyter-core
prettytable==3.6.0
    # via hyperspy
prometheus-client==0.15.0
    # via jupyter-server
prompt-toolkit==3.0.36
    # via ipython
psutil==5.9.4
    # via
    #   ipykernel
    #   ipyparallel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pyelli==0.15.0
    # via -r requirements.in
pygments==2.14.0
    # via
    #   ipython
    #   nbconvert
pyparsing==3.0.9
    # via matplotlib
pyrsistent==0.19.3
    # via jsonschema
pytest-runner==6.0.0
    # via h5py-wrapper
python-dateutil==2.8.2
    # via
    #   hyperspy
    #   ipyparallel
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2022.7.1
    # via
    #   nexusutils
    #   pandas
pywavelets==1.4.1
    # via scikit-image
pyyaml==6.0
    # via
    #   dask
    #   hdfdict
    #   hyperspy
    #   nexusutils
    #   pyelli
pyzmq==25.0.0
    # via
    #   ipykernel
    #   ipyparallel
    #   jupyter-client
    #   jupyter-server
rapidfuzz==2.13.7
    # via pyelli
requests==2.28.2
    # via hyperspy
scikit-image==0.19.3
    # via hyperspy
scipy==1.10.0
    # via
    #   ase
    #   hyperspy
    #   pyelli
    #   scikit-image
    #   sparse
send2trash==1.8.0
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
sparse==0.13.0
    # via hyperspy
stack-data==0.6.2
    # via ipython
sympy==1.11.1
    # via hyperspy
terminado==0.17.1
    # via jupyter-server
tifffile==2022.10.10
    # via
    #   h5grove
    #   hyperspy
    #   scikit-image
tinycss2==1.2.1
    # via nbconvert
toolz==0.12.0
    # via
    #   dask
    #   hyperspy
    #   partd
tornado==6.2
    # via
    #   ipykernel
    #   ipyparallel
    #   jupyter-client
    #   jupyter-server
    #   terminado
tqdm==4.64.1
    # via
    #   hyperspy
    #   ipyparallel
traitlets==5.8.1
    # via
    #   comm
    #   ipykernel
    #   ipyparallel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
traits==6.4.1
    # via hyperspy
urllib3==1.26.14
    # via requests
wcwidth==0.2.6
    # via
    #   prettytable
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.4.2
    # via jupyter-server
xarray==2023.1.0
    # via nexusutils
zarr==2.13.6
    # via hyperspy
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
